MMT-3936: Moving MMT_COOKIE to sharedConstants
